requirements.txt
boto3==1.28.37
python-dotenv===1.0.0